+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docker building script for whole components [23](https://github.com/ncsa/standalone-smm-analytics/issues/23)
 - Docker compose launch script [45](https://github.com/ncsa/standalone-smm-analytics/issues/45)
 - Docker compose file using traefik [46](https://github.com/ncsa/standalone-smm-analytics/issues/46)
+- Environment variables for turn on off twitter and reddit [73](https://github.com/ncsa/standalone-smm-analytics/issues/73)
 
 ### Changed
 - Hard coded rabbimq url changed to env variable [18](https://github.com/ncsa/standalone-smm-analytics/issues/18)